+++ src/shader/slang/slang_emit.c
@@ -537,7 +537,7 @@ emit_arith(slang_emit_info *emitInfo, slang_ir_node *n)
     /* result storage */
     if (!n->Store) {
 	GLint size = n->Children[0]->Store
-		     ? n->Children[0]->Store->Size : info->ResultSize;
+		     ? n->Children[0]->Store->Size : (GLint)info->ResultSize;
 	if (!alloc_temp_storage(emitInfo, n, size))
 	    return NULL;
     }
